Make flow capture work when mitm is in "upstream" mode.
When we use an upstream proxy, it does not populate the server ip address after connecting (see tunnel.py). So being connected does not imply we have a value for the server ip address, instead just set the serverIPAddress if we have a value for it.

User peername instead of ip_address because ip_address is deprecated.